feat: move utils to new logger
