--- src/utils/dimensionsPropType.ts
@@ -1,6 +1,0 @@
-import PropTypes from 'prop-types';
-
-export default PropTypes.shape({
-  height: PropTypes.number,
-  width: PropTypes.number,
-});
