--- src/utils/directionPropType.ts
@@ -1,3 +1,0 @@
-import PropTypes from 'prop-types';
-
-export default PropTypes.oneOf(['bottom', 'left', 'right', 'top']);
