+++ src/utils/platform.ts
@@ -1,2 +1,3 @@
-export const cloudPlatform = ios ? 'iCloud' : 'Google Drive';
-export const cloudPlatformAccountName = ios ? 'Apple iCloud' : 'Google';
+import { IS_IOS } from '@/env';
+export const cloudPlatform = IS_IOS ? 'iCloud' : 'Google Drive';
+export const cloudPlatformAccountName = IS_IOS ? 'Apple iCloud' : 'Google';
